Update README.md
KPI's Requirement

The HR department is responsible for monitoring and managing various aspects of employee data to ensure the organization maintains a healthy workforce. However, there is a lack of clear performance indicators to track and analyze key HR metrics. Therefore, there is a need to design and implement a set of KPIs to address the following points:

1. Employee Count:

The HR department lacks visibility into the total number of employees, making it challenging to assess workforce size and plan for future growth or downsizing effectively.

2. Attrition Count:

The organization lacks a standardized method to track employee attrition, resulting in incomplete and unreliable data on the number of employees who have left the organization.

3. Attrition Rate:

Without a clear measure of attrition rate, the organization cannot assess the overall turnover level or compare it with

industry benchmarks, hindering the ability to gauge employee satisfaction and engagement.

4. Active Employees:

The organization lacks a mechanism to differentiate between active and inactive employees, leading to difficulties in accurately assessing the current workforce's productivity and capacity.

5. Average Age:

The HR department lacks visibility into the average age of employees, making it difficult to evaluate workforce demographics, succession planning, and the organization's ability to attract and retain younger talent.